[buteo-syncml] Fix profile removal - set absolute file path when calling setDatabaseFilePath()
msyncd was failing to remove sync profiles when requested as this
fails if the plugin does not cleanUp() successfully. This was a problem
as SyncAgent::cleanUp() always failed as it was locating the database
file with a non-absolute path.

The code for creating the absolute databse file path was previously in
DatabaseHandler; this patch moves the code to SyncAgentConfig to ensure
that it sets the database file path to an absolute file path as
expected.